requirements.txt
pillow
requests==2.32.5
beautifulsoup4==4.14.2
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.2
cryptography==46.0.3
python-dotenv==1.1.1
markdown==3.9
selenium==4.37.0
pillow==12.0.0
openai==2.6.0